backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
django-cors-headers==4.3.0